adjust after moving objects
