+++ src/neptune/metadata_containers/metadata_container.py
@@ -94,9 +94,9 @@ from neptune.metadata_containers.abstract import (
     NeptuneObjectCallback,
 )
 from neptune.metadata_containers.metadata_containers_table import Table
-from neptune.progress_bar import ProgressBarCallback
 from neptune.types.mode import Mode
 from neptune.types.type_casting import cast_value
+from neptune.typing import ProgressBarCallback
 from neptune.utils import stop_synchronization_callback
 
 if TYPE_CHECKING:
