+++ src/neptune/metadata_containers/project.py
@@ -48,8 +48,8 @@ from neptune.metadata_containers import MetadataContainer
 from neptune.metadata_containers.abstract import NeptuneObjectCallback
 from neptune.metadata_containers.metadata_containers_table import Table
 from neptune.metadata_containers.utils import prepare_nql_query
-from neptune.progress_bar import ProgressBarCallback
 from neptune.types.mode import Mode
+from neptune.typing import ProgressBarCallback
 
 
 class Project(MetadataContainer):
